Finalização do Readme, limpeza de código
